removed meetings and optimized
