+++ backend/get_schedule/lambda_function.py
@@ -3,7 +3,9 @@ from dotenv import load_dotenv
 import os
 import itertools
 import json
-from math import pi, sqrt, cos
+
+from section_filters import fetch_section_data, delete_unpreferred_sections, delete_closed_sections, check_for_no_section
+from schedule_score import compute_schedule_score
 
 load_dotenv()
 
@@ -14,13 +16,6 @@ db = client[os.environ.get("MONGODB_DB")]
 classes = db["classes"]
 locations = db["locations"]
 
-weights = {
-    "distance": 25,
-    "back_to_back": 15,
-    "time": 50,
-    "section": 10,
-    "lunch": 7.5
-}
 
 class ScheduleException(Exception):
     pass
@@ -28,62 +23,6 @@ class ScheduleException(Exception):
 class ScheduleOverlapException(Exception):
     # When a schedule is not possible because 2 classes overlap time-wise
     pass
-
-def check_for_no_section(classes_to_take, open_sections_filter_applied):
-    for clas in classes_to_take:
-        if len(clas["sections"]) == 0:
-            if (open_sections_filter_applied):
-                raise ScheduleException(f"No open sections found for {clas['code']} {clas['number']} in your specified time range. Please try again with a different time range.")
-            else:
-                raise ScheduleException(f"No sections found for {clas['code']} {clas['number']} in your specified time range. Please try again with a different time range.")
-
-def fetch_section_data(class_list):
-    """
-    Fetches the section data from the database
-    """
-    for class_ in class_list:
-        clas_mongo_object = classes.find_one({"code": class_["code"], "number": class_[
-                                             "number"], "year": 2024, "semester": "spring"})
-        class_["sections"] = []
-        for section in clas_mongo_object["sections"]:
-            if section["crn"] in class_["crn_list"]:
-
-                if section["type_code"] == "OLC":
-                    continue
-
-                section["class"] = {
-                    "code": class_["code"],
-                    "number": class_["number"]
-                }
-                
-                class_["sections"].append(section)
-
-    return class_list
-def is_preferred_time(section, start_time, end_time):
-    earliest_meeting = start_time
-    latest_meeting = end_time
-
-    if section['start_time'] is None or section['end_time'] is None:
-        return True
-    if section['start_time'].hour < earliest_meeting:
-        earliest_meeting = section['start_time'].hour
-    if section['end_time'].hour > latest_meeting:
-        latest_meeting = section['end_time'].hour
-
-    return earliest_meeting >= start_time and latest_meeting <= end_time
-
-
-def delete_unpreferred_sections(class_list, start_time, end_time):
-    for clas in class_list:
-        clas["sections"] = [section for section in clas['sections'] if is_preferred_time(section, start_time, end_time)]
-    return class_list
-
-
-def delete_closed_sections(class_list, open_sections_only):
-    if open_sections_only:
-        for clas in class_list:
-            clas["sections"] = [section for section in clas['sections'] if section['enrollment_status'].lower() != 'closed']
-    return class_list
 
 def apply_hard_filters(user_preferences):
     class_list = fetch_section_data(user_preferences["classes"])
@@ -135,152 +74,6 @@ def generate_schedule_combinations(class_list, user_preferences):
     return possible_schedules
 
 
-def convert_to_time_based(list_of_sections):
-    schedule_slots = {
-        'W': [None]*23,
-        'R': [None]*23,
-        'F': [None]*23,
-        'T': [None]*23,
-        'M': [None]*23,
-    }
-
-    for section in list_of_sections:
-        if section['start_time'] is None or section['end_time'] is None:
-            continue
-
-        start_hour = section['start_time'].hour
-        end_hour = section['end_time'].hour
-
-        for day in section['days'].strip():
-            for hour in range(start_hour, end_hour+1):
-                if schedule_slots[day][hour] is not None:
-                    raise ScheduleOverlapException(f"Schedule overlap detected for {section['class']['code']} {section['class']['number']} and {schedule_slots[day][hour]['class']['code']} {schedule_slots[day][hour]['class']['number']}")
-                schedule_slots[day][hour] = section
-
-    return schedule_slots
-
-
-def pythagorian_distance(coords1, coords2):
-    lat1 = coords1[0]  
-    lon1 = coords1[1]  
-    lat2 = coords2[0] 
-    lon2 = coords1[1] 
-
-    avgLatDeg = (lat1 + lat2) / 2
-    avgLat = avgLatDeg * (pi/180)
-
-    d_ew = (lon2 - lon1) * cos(avgLat)
-    d_ns = (lat2 - lat1)
-    approxDegreesSq = (d_ew * d_ew) + (d_ns * d_ns)
-    d_degrees = sqrt(approxDegreesSq)
-    EarthAvgMeterPerGreatCircleDegree = (6371000 * 2 * pi) / 360
-    return d_degrees * EarthAvgMeterPerGreatCircleDegree
-
-def list_of_successive_distances(schedule):
-
-    list_of_distances = []
-    for day in schedule:
-        for hour in range(0, 23):
-            if schedule[day][hour] is not None and schedule[day][hour + 1] is not None:
-                coords1 = schedule[day][hour]['coordinates']
-                coords2 = schedule[day][hour + 1]['coordinates']
-                if coords1 is not None and coords2 is not None:
-                    list_of_distances.append(pythagorian_distance(coords1, coords2))
-
-    return [x for x in list_of_distances if x != 0.0]
-
-
-
-def distance_score(list_of_sections, max_distance) -> float:
-    # Returns a score between -distance_weight and distance_weight based on how far apart the classes are
-    list_of_distances = list_of_successive_distances(list_of_sections)
-
-    if len(list_of_distances) == 0:
-        return 0
-
-    mean_distance = sum(list_of_distances) / len(list_of_distances)
-
-    # Find the distance score
-    if mean_distance > max_distance:
-        percent_over = (mean_distance - max_distance) / max_distance
-        return -(percent_over * weights["distance"])
-    elif mean_distance < max_distance:
-        percent_under = (max_distance - mean_distance) / max_distance
-        return percent_under * weights["distance"]
-    else:
-        return 0
-
-
-def back_to_back_score(list_of_sections, back_to_back) -> float:
-    back_to_back_count = 0
-    total_count = 0
-
-    for day in list_of_sections:
-        for hour in range(0, 23):
-            if list_of_sections[day][hour] is not None and list_of_sections[day][hour + 1] is not None:
-                total_count += 1
-                if list_of_sections[day][hour]['crn'] != list_of_sections[day][hour + 1]['crn']:
-                    back_to_back_count += 1
-
-    if total_count == 0:
-        return 0
-    elif back_to_back:
-        return (back_to_back_count / total_count) * weights["back_to_back"]
-    else:
-        return (-back_to_back_count / total_count) * weights["back_to_back"]
-
-
-def time_score(list_of_sections, pref_time):
-    time_list = []
-
-    for day in list_of_sections:
-        for hour in range(0, 23):
-            if list_of_sections[day][hour] is not None:
-                time_list.append(hour)
-
-    if len(time_list) == 0:
-        return 0
-    else:
-        mean_time = sum(time_list) / len(time_list)
-        return (1 - (abs(mean_time - pref_time) / pref_time)) * weights["time"]
-
-
-def section_score(list_of_sections, pref_sections):
-    section_count = 0
-    for day in list_of_sections:
-        for hour in range(0, 23):
-            if list_of_sections[day][hour] is not None and list_of_sections[day][hour]['crn'] in pref_sections:
-                section_count += 1
-
-    return section_count * weights["section"]
-
-
-def lunch_score(list_of_sections, lunch):
-    lunch_count = 0
-    for day in list_of_sections:
-        for hour in range(0, 23):
-            if list_of_sections[day][hour] is None and hour >= lunch['start'] and hour < lunch['end']:
-                duration_hours = [list_of_sections[day][x] is None for x in range(hour, hour + lunch['duration']) if x < 23]
-                if all(duration_hours):
-                    lunch_count += 1
-                    break
-    return (lunch_count - 5) * weights["lunch"]
-
-def compute_schedule_score(list_of_sections, user_preferences) -> float:
-    schedule = convert_to_time_based(list_of_sections)
-    
-    if user_preferences['max_distance'] > 10_000:
-        ds = 0
-    else :
-        ds = distance_score(schedule, user_preferences["max_distance"])
-
-    btb = back_to_back_score(schedule, user_preferences["back_to_back"])
-    ts = time_score(schedule, user_preferences["pref_time"])
-    ss = section_score(schedule, user_preferences["pref_sections"])
-    ls = lunch_score(schedule, user_preferences["lunch"])
-    return (ds + btb + ts + ss + ls), schedule
-
-
 def sort_schedules(possible_schedules, user_preferences):
     sorted_schedules = []
     for schedule in possible_schedules:
@@ -301,6 +94,14 @@ def sort_schedules(possible_schedules, user_preferences):
     print(f"Number of possible schedules: {len(sorted_schedules)}")
     return sorted_schedules
 
+
+def is_full_semester_schedule(class_list):
+    """
+    Return true if 
+    - every class in the list is a full semester class
+    - all the half semester classes are in the same half semester
+    """
+    half_semester_classes = []
 
 def get_schedule(user_preferences):
     """
