+++ backend/get_schedule/requirements.txt
@@ -1,3 +1,2 @@
 pymongo
 python-dotenv
-haversine
